+++ src/common/enums/index.ts
@@ -1,0 +1,8 @@
+export {
+  GenreEnum,
+  RoleEnum,
+  StatusEnum,
+  pgGenreEnum,
+  pgRoleEnum,
+  pgStatusEnum
+} from './db-enums'
