+++ src/common/lib/index.ts
@@ -1,1 +1,2 @@
 export { createCustomId } from './id'
+export { compareEncryptValue, encryptData } from './bcrypt'
