+++ src/drizzle/schema/employees.ts
@@ -1,12 +1,10 @@
 import { pgTable, varchar, text } from 'drizzle-orm/pg-core'
 import { relations } from 'drizzle-orm'
 
-import { users, companies } from '.'
+import { users, companies, pgStatusEnum } from '.'
 
 import { createCustomId } from 'src/common/lib'
 import { timestamps } from 'src/common/utils'
-
-import { pgStatusEnum } from 'src/enums'
 
 export const employees = pgTable('employees', {
   id: text('id')
