+++ src/drizzle/schema/roles.ts
@@ -1,12 +1,10 @@
 import { pgTable, text } from 'drizzle-orm/pg-core'
 import { relations } from 'drizzle-orm'
 
-import { users } from '.'
+import { pgRoleEnum, users } from '.'
 
 import { timestamps } from 'src/common/utils'
 import { createCustomId } from 'src/common/lib'
-
-import { pgRoleEnum } from 'src/enums'
 
 export const roles = pgTable('roles', {
   id: text('id')
