+++ src/drizzle/schema/rooms.ts
@@ -1,11 +1,10 @@
 import { pgTable, varchar, text } from 'drizzle-orm/pg-core'
 import { relations } from 'drizzle-orm'
 
-import { units } from '.'
+import { units, pgStatusEnum } from '.'
 
 import { timestamps } from 'src/common/utils'
 import { createCustomId } from 'src/common/lib'
-import { pgStatusEnum } from 'src/enums'
 
 export const rooms = pgTable('rooms', {
   id: text('id')
