--- src/enums/index.ts
@@ -1,3 +1,0 @@
-export { GenreEnum, RoleEnum, StatusEnum } from './db-enums'
-
-export { pgGenreEnum, pgRoleEnum, pgStatusEnum } from './db-enums'
